appended timestamps move to top of post
developed a standard timestamp format according to what was already in place. js takes that when appended to the end of a post, makes it invisible, replaces the built-in time stamp at the top of the post with the user added one. Need to add code to make it render as time from posting instead of just the time posted.